requirements.txt
fastapi==0.110.0
h5py==3.11.0
Jinja2==3.1.3
joblib==1.3.2
keras==3.2.1
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
torch==2.2.2
torchvision==0.17.2
ultralytics==8.1.47
urllib3==2.2.1
uvicorn==0.29.0
